+++ lib/sequenced.rb
@@ -1,4 +1,3 @@
-require 'sequenced/exceptions'
 require 'sequenced/acts_as_sequenced'
 
 ActiveRecord::Base.send(:include, Sequenced::ActsAsSequenced)