+++ lib/sequenced/exceptions.rb
@@ -1,4 +1,0 @@
-module Sequenced
-  class SequencedError < RuntimeError; end
-  class InvalidAttributeError < SequencedError; end
-end